fix server action problem
